+++ akka-http-core/src/main/scala/akka/http/impl/engine/http2/Http2.scala
@@ -9,12 +9,16 @@ import akka.annotation.InternalApi
 import akka.dispatch.ExecutionContexts
 import akka.event.LoggingAdapter
 import akka.http.impl.engine.server.{ MasterServerTerminator, UpgradeToOtherProtocolResponseHeader }
+import akka.http.impl.util.LogByteStringTools
+import akka.http.scaladsl.Http.OutgoingConnection
 import akka.http.scaladsl.{ ConnectionContext, Http, HttpsConnectionContext }
 import akka.http.scaladsl.Http.ServerBinding
 import akka.http.scaladsl.model._
 import akka.http.scaladsl.model.headers.{ Connection, RawHeader, Upgrade, UpgradeProtocol }
 import akka.http.scaladsl.model.http2.Http2SettingsHeader
+import akka.http.scaladsl.settings.ClientConnectionSettings
 import akka.http.scaladsl.settings.ServerSettings
+import akka.stream.TLSClosing
 import akka.stream.TLSProtocol.{ SslTlsInbound, SslTlsOutbound }
 import akka.stream.impl.io.TlsUtils
 import akka.stream.scaladsl.{ Flow, Keep, Sink, Source, TLS, TLSPlacebo, Tcp }
@@ -183,6 +187,38 @@ private[http] final class Http2Ext(private val config: Config)(implicit val syst
 
     ProtocolSwitch(_ => getChosenProtocol(), http1, http2) join
       tls
+  }
+
+  def outgoingConnection(host: String, port: Int, connectionContext: HttpsConnectionContext, clientConnectionSettings: ClientConnectionSettings, log: LoggingAdapter): Flow[HttpRequest, HttpResponse, Future[OutgoingConnection]] = {
+    def createEngine(): SSLEngine = {
+      val engine = connectionContext.sslContextData match {
+        // TODO FIXME configure hostname verification for this case
+        case Left(ssl) =>
+          val e = ssl.sslContext.createSSLEngine(host, port)
+          TlsUtils.applySessionParameters(e, ssl.firstSession)
+          e
+        case Right(e) => e(Some((host, port)))
+      }
+      engine.setUseClientMode(true)
+      Http2AlpnSupport.clientSetApplicationProtocols(engine, Array("h2"))
+      engine
+    }
+
+    val stack = Http2Blueprint.clientStack(clientConnectionSettings, log) atop
+      Http2Blueprint.unwrapTls atop
+      LogByteStringTools.logTLSBidiBySetting("client-plain-text", clientConnectionSettings.logUnencryptedNetworkBytes) atop
+      TLS(createEngine _, closing = TLSClosing.eagerClose)
+
+    stack.joinMat(clientConnectionSettings.transport.connectTo(host, port, clientConnectionSettings)(system.classicSystem))(Keep.right)
+  }
+
+  def outgoingConnectionPriorKnowledge(host: String, port: Int, clientConnectionSettings: ClientConnectionSettings, log: LoggingAdapter): Flow[HttpRequest, HttpResponse, Future[OutgoingConnection]] = {
+    val stack = Http2Blueprint.clientStack(clientConnectionSettings, log) atop
+      Http2Blueprint.unwrapTls atop
+      LogByteStringTools.logTLSBidiBySetting("client-plain-text", clientConnectionSettings.logUnencryptedNetworkBytes) atop
+      TLSPlacebo()
+
+    stack.joinMat(clientConnectionSettings.transport.connectTo(host, port, clientConnectionSettings)(system.classicSystem))(Keep.right)
   }
 
 }
