+++ akka-http-core/src/main/scala/akka/http/javadsl/model/RequestResponseAssociation.scala
@@ -14,12 +14,12 @@ import scala.concurrent.Promise
  * A marker trait for attribute values that should be (automatically) carried over from request to response.
  */
 @ApiMayChange
-trait RequestResponseAssociation extends akka.http.scaladsl.model.RequestResponseAssociation
+trait RequestResponseAssociation
 
 /**
  * An association for completing a future when the response arrives.
  */
-final class ResponseFuture(val future: CompletableFuture[HttpResponse]) extends RequestResponseAssociation
+final class ResponseFuture(val future: CompletableFuture[HttpResponse]) extends akka.http.scaladsl.model.RequestResponseAssociation
 object ResponseFuture {
   val KEY = AttributeKey.create[ResponseFuture]("association-future-handle", classOf[ResponseFuture])
   def apply(promise: CompletableFuture[HttpResponse]): ResponseFuture = new ResponseFuture(promise)
