+++ akka-http-core/src/main/scala/akka/http/scaladsl/Http.scala
@@ -392,8 +392,6 @@ class HttpExt private[http] (private val config: Config)(implicit val system: Ex
    * is no pooling and you are yourself responsible for lifecycle management of the connection. For a more convenient
    * Request level API see [[singleRequest()]]
    *
-   * The builder defaults to build a connection using HTTP/1.1 over a plaintext connection to port 80.
-   *
    * @return A builder to configure more specific setup for the connection and then build a `Flow[Request, Response, Future[OutgoingConnection]]`.
    */
   def connectionTo(host: String): OutgoingConnectionBuilder = OutgoingConnectionBuilderImpl(host, system)
