+++ akka-http-core/src/main/scala/akka/http/scaladsl/OutgoingConnectionBuilder.scala
@@ -18,7 +18,7 @@ import scala.concurrent.Future
 
 /**
  * Builder for setting up a flow that will create one single connection per materialization to the specified host.
- * When customization is done, the flow is created using [[#connectionFlow()]].
+ * When customization is done, the flow is created using [[#http()]], [[#https()]], [[#http2()]] or [[#http2WithPriorKnowledge()]].
  *
  * Not for user extension
  */
