+++ akka-http-core/src/main/scala/akka/http/scaladsl/model/RequestResponseAssociation.scala
@@ -11,20 +11,20 @@ import scala.concurrent.Promise
  * A marker trait for attribute values that should be (automatically) carried over from request to response.
  */
 @ApiMayChange
-trait RequestResponseAssociation
+trait RequestResponseAssociation extends akka.http.javadsl.model.RequestResponseAssociation
 
 /**
  * A simple value holder class implementing RequestResponseAssociation.
  */
 @ApiMayChange
-case class SimpleRequestResponseAttribute[T](value: T) extends RequestResponseAssociation
+final case class SimpleRequestResponseAttribute[T](value: T) extends RequestResponseAssociation
 
 /**
  * An association for completing a future when the response arrives.
  */
-final class ResponseFuture(val promise: Promise[HttpResponse]) extends RequestResponseAssociation
-object ResponseFuture {
-  val Key = AttributeKey[ResponseFuture]("association-future-handle")
-  def apply(promise: Promise[HttpResponse]): ResponseFuture = new ResponseFuture(promise)
+final class ResponsePromise(val promise: Promise[HttpResponse]) extends RequestResponseAssociation
+object ResponsePromise {
+  val Key = AttributeKey[ResponsePromise]("association-future-handle")
+  def apply(promise: Promise[HttpResponse]): ResponsePromise = new ResponsePromise(promise)
 }
